+++ DTOs/Response/AuthResponseDTOs/LoginResponseDto.cs
@@ -3,6 +3,7 @@
     public class LoginResponseDto
     {
         public string Email { get; set; } = default!;
+        public string Nombre { get; set; } = default!;
         public string Token { get; set; } = default!;
     }
 }
