+++ Repositories/Interfaces/IUsuarioRepository.cs
@@ -9,8 +9,6 @@ namespace template_csharp_dotnet.Repositories.Interfaces
         Task<Usuario> DeleteUserWithRelations(int id);
         Task<List<Usuario>> GetAllUsersWithRelations();
         Task<Usuario> GetUserByEmail(string email);
-        string HashPassword(Usuario usuario, string providedPassword);
-        PasswordVerificationResult VerifyPassword(Usuario usuario, string providedPassword);
         Task<Usuario> AuthenticateUser(string email, string password);
     }
 }
